gun rate able to upgrade
also change the start menu with finalgradesystem as a child of difficulty manager

+++ FPS-Prototype/Assets/Scripts/Player/PlayerAbilities.cs
@@ -11,7 +11,7 @@ public class PlayerAbilities : MonoBehaviour
     //pistol
     public int w1DmgMod;
     public int w1SpeedMod;
-    public int w1RateMod;
+    public float w1RateMod;
     public bool w1Major;
     //bow
     public int w2DmgMod;
